+++ Source/ACE.Server/Network/GameEvent/Events/GameEventResetTrade.cs
@@ -1,0 +1,13 @@
+using ACE.Entity;
+
+namespace ACE.Server.Network.GameEvent.Events
+{
+    public class GameEventResetTrade : GameEventMessage
+    {
+        public GameEventResetTrade(Session session, ObjectGuid whoReset)
+            : base(GameEventType.ResetTrade, GameMessageGroup.UIQueue, session)
+        {
+            Writer.WriteGuid(whoReset);
+        }
+    }
+}